+++ EroojaNetwork/EroojaNetwork/APIRequest/EroojaAPIRequest.swift
@@ -67,67 +67,4 @@ public class EroojaAPIRequest {
             }
         })
     }
-    
-    public func requestNicknameExist(nickname: String, token: String, completion: @escaping (Result<Bool, EroojaAPIError>) -> Void) {
-        let urlString = UserAPIRequest.RequestType.nicknameExist(nickname).requestURL.absoluteString
-        let parameters = UserAPIRequest.RequestType.nicknameExist(nickname).requestParameter
-        
-        let headers: HTTPHeaders = ["Authorization" : "Bearer \(token)"]
-        
-        AF.request(urlString, method: .post, parameters: parameters, encoder: JSONParameterEncoder.default, headers: headers).responseString(completionHandler: { response in
-            switch response.result {
-            case .success(_):
-                ELog.debug("[EroojaAPIRequest]: Check Nickname Exist ")
-                ELog.debug(response.value!)
-                if response.value == "false" {
-                    completion(.success(false))
-                } else if response.value == "true" {
-                    completion(.success(true))
-                } else {
-                    completion(.failure(.decodeError))
-                }
-            case .failure(_):
-                completion(.failure(.urlRequestError))
-            }
-        })
-    }
-    
-    public func requestNicknameUpdate(nickname: String, token: String, completion: @escaping (Result<NSDictionary, EroojaAPIError>) -> Void) {
-        let urlString = UserAPIRequest.RequestType.nicknameUpdate(nickname).requestURL.absoluteString
-        let headers: HTTPHeaders = ["Authorization" : "Bearer \(token)"]
-        let parameters = UserAPIRequest.RequestType.nicknameUpdate(nickname).requestParameter
-        
-        AF.request(urlString, method: .put, parameters: parameters, encoder: JSONParameterEncoder.default, headers: headers).responseJSON(completionHandler: { response in
-            switch response.result {
-            case .success(_):
-                if let responseValue = (response.value as? NSDictionary) {
-                    completion(.success(responseValue))
-                } else {
-                    completion(.failure(.decodeError))
-                }
-            case .failure(_):
-                completion(.failure(.urlRequestError))
-            }
-        })
-    }
-    
-    public func requestUploadImage(imageData: Data, token: String, completion: @escaping (Result<NSDictionary, EroojaAPIError>) -> Void) {
-        let urlString = UserAPIRequest.RequestType.profileImageUpdate.requestURL
-        let headers: HTTPHeaders = ["Authorization" : "Bearer \(token)", "Content-Type":"multipart/form-data"]
-        
-        AF.upload(multipartFormData: { multipartFormData in
-            multipartFormData.append(imageData, withName: "multipartImageFile", fileName: "file.jpg", mimeType: "image/jpg")
-        }, to: urlString, method: .put, headers: headers).responseJSON(completionHandler: { response in
-            switch response.result {
-            case .success(_):
-                if let responseValue = (response.value as? NSDictionary) {
-                    completion(.success(responseValue))
-                } else {
-                    completion(.failure(.decodeError))
-                }
-            case .failure(_):
-                completion(.failure(.urlRequestError))
-            }
-        })
-    }
 }
